+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `trim` method
 - Add `replaceBetween` method
 - Add `removeNewLine` method
+- Add `capitalizeEverySentence` method
 
 ## 0.0.1 (2024-09-02) - Not for Production